FNIRT_single: Added inverse FLIRT, code cleanup and required ${1}
Added inverse FLIRT which I overlooked, note that A*B=B^-1*A^-1 so the
resulting inverse FLIRT is applied in the as a postmat, as opposed to
the "premat" of the initial FLIRT.

Some code cleanup was done ("echo ." --> "echo").

Additionally added the requirement of adding a subject name as an
argument when calling the FNIRT_single shell script.